added chat

+++ rentit/src/main/java/kg/neobis/rentit/service/BookingService.java
@@ -3,24 +3,20 @@ package kg.neobis.rentit.service;
 import kg.neobis.rentit.dto.BookingRegistrationDto;
 import kg.neobis.rentit.dto.BookingRequestDto;
 import kg.neobis.rentit.dto.UserBookingDto;
-import kg.neobis.rentit.entity.Booking;
 import kg.neobis.rentit.entity.Calendar;
-import kg.neobis.rentit.entity.ImageProduct;
-import kg.neobis.rentit.entity.Product;
-import kg.neobis.rentit.entity.User;
+import kg.neobis.rentit.entity.*;
 import kg.neobis.rentit.enums.BookingStatus;
 import kg.neobis.rentit.exception.AlreadyExistException;
 import kg.neobis.rentit.exception.BadRequestException;
+import kg.neobis.rentit.exception.ProductViolationException;
 import kg.neobis.rentit.exception.ResourceNotFoundException;
-import kg.neobis.rentit.repository.BookingRepository;
-import kg.neobis.rentit.repository.CalendarRepository;
-import kg.neobis.rentit.repository.ProductRepository;
-import kg.neobis.rentit.repository.UserRepository;
+import kg.neobis.rentit.repository.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.YearMonth;
@@ -39,11 +35,17 @@ public class BookingService {
 
     private final CalendarRepository calendarRepository;
 
+    private final ImageProductRepository imageProductRepository;
+
+    private final ComplaintRepository complaintRepository;
+
+    private final ReviewRepository reviewRepository;
+
 
     public List<UserBookingDto> getUserBookings() {
         User user = getAuthentication();
 
-        return user.getBookings().stream()
+        return user == null ? null : user.getBookings().stream()
                 .map(
                         entity -> {
                             UserBookingDto dto = new UserBookingDto();
@@ -51,18 +53,11 @@ public class BookingService {
                             dto.setBookingId(entity.getId());
                             dto.setStatus(entity.getBookingStatus().getStatus());
                             dto.setProductId(entity.getProduct().getId());
-                            dto.setPrice(entity.getProduct().getPrice());
+                            dto.setTotalPrice(entity.getTotalPrice());
                             dto.setProductTitle(entity.getProduct().getTitle());
-
-                            List<ImageProduct> imageProductList = entity.getProduct().getImageProduct();
-
-                            for (ImageProduct imageProduct : imageProductList) {
-
-                                if (imageProduct.getOrderNumber() == 0) {
-                                    dto.setMainImageUrl(imageProduct.getImage().getUrl());
-                                    break;
-                                }
-                            }
+                            dto.setMainImageUrl(returnMainImageUrl(entity));
+                            dto.setBookFrom(entity.getDateFrom());
+                            dto.setBookTill(entity.getDateTill());
 
                             return dto;
                         }
@@ -88,42 +83,51 @@ public class BookingService {
                     dto.setBookingId(entity.getId());
                     dto.setClientId(entity.getUser().getId());
                     dto.setClientName(entity.getUser().getLastName() + " " + entity.getUser().getFirstName());
-                    dto.setPrice(entity.getProduct().getPrice());
+                    dto.setTotalPrice(entity.getTotalPrice());
                     dto.setProductTitle(entity.getProduct().getTitle());
                     dto.setProductId(entity.getProduct().getId());
+                    dto.setBookDateFrom(entity.getDateFrom());
+                    dto.setBookDateTill(entity.getDateTill());
+                    dto.setMainImageUrl(returnMainImageUrl(entity));
 
-                    List<ImageProduct> imageProductList = entity.getProduct().getImageProduct();
+                    int rating = 100 - complaintRepository
+                            .getComplaintsNumberByAddresseeId(dto.getClientId());
 
-                    for (ImageProduct imageProduct : imageProductList) {
+                    for(Product p: entity.getUser().getProducts()) {
+                        List<Review> reviews = reviewRepository.findAllByProductId(p.getId());
 
-                        if (imageProduct.getOrderNumber() == 0) {
-                            dto.setMainImageUrl(imageProduct.getImage().getUrl());
-                            break;
+                        double sum = (double) reviews.stream()
+                                .mapToInt(Review::getStar).sum() / reviews.size();
+
+                        if(sum >= 4.9) {
+                            rating += 5;
                         }
                     }
+
+                    dto.setProfileRating(rating > 100 ? 100 : Math.max(rating, 1));
 
                     return dto;
                 })
                 .collect(Collectors.toList());
     }
 
-    public BookingRegistrationDto bookProduct(BookingRegistrationDto dto) {
-        LocalDate from = dto.getDateFrom();
-        LocalDate till = dto.getDateTill();
+    private String returnMainImageUrl(Booking entity) {
+        ImageProduct imageProduct =
+                imageProductRepository.findByProductIdAndOrderNumber(entity.getProduct().getId(),
+                        (byte) 1);
 
-        for (Booking entity : bookingRepository.findAll()) {
-            LocalDate efrom = entity.getDateFrom();
-            LocalDate etill = entity.getDateTill();
-            if(entity.getBookingStatus().equals(BookingStatus.ACCEPTED)) {
-                if ((from.compareTo(efrom) >= 0 && from.compareTo(etill) < 0) ||
-                        (from.compareTo(efrom) > 0 && from.compareTo(etill) <= 0) ||
-                        (till.compareTo(efrom) >= 0 && till.compareTo(etill) < 0) ||
-                        (till.compareTo(efrom) > 0 && till.compareTo(etill) <= 0)) {
-                    throw new AlreadyExistException("The given dates are already booked.");
-                }
+        if (imageProduct != null) {
+            Image image = imageProduct.getImage();
+            if (image != null && image.getUrl().startsWith("http")) {
+                return image.getUrl().replace("http", "https");
             }
         }
 
+        return "";
+    }
+
+    @Transactional
+    public BookingRegistrationDto bookProduct(BookingRegistrationDto dto) {
         User user = getAuthentication();
 
         if (user == null) {
@@ -135,8 +139,35 @@ public class BookingService {
                         () -> new ResourceNotFoundException("Product was not found with ID: " + dto.getProductId())
                 );
 
-        if(user.getId().equals(product.getUser().getId())) {
-            throw new BadRequestException("You cannot add your product to favorites.");
+        LocalDate sumFrom = dto.getDateFrom();
+        LocalDate sumTill = dto.getDateTill();
+
+        int days = 1;
+
+        while (!sumFrom.isEqual(sumTill)) {
+            sumFrom = sumFrom.plusDays(1);
+            days++;
+        }
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(dto.getDateFrom(),
+                dto.getDateTill(), product.getId());
+
+        if (days != calendar.size()) {
+            throw new BadRequestException("Incorrect date input. Some of the given dates are not eligible" +
+                    " for booking.");
+        }
+
+        for (Calendar entity : calendar) {
+            if (entity.isBooked()) {
+                throw new AlreadyExistException("The date is already booked: " + entity.getDate());
+            } else {
+                entity.setBooked(true);
+            }
+            calendarRepository.save(entity);
+        }
+
+        if (user.getId().equals(product.getUser().getId())) {
+            throw new BadRequestException("You cannot book your product.");
         }
 
         Booking booking = new Booking();
@@ -146,8 +177,8 @@ public class BookingService {
         booking.setDateFrom(dto.getDateFrom());
         booking.setDateTill(dto.getDateTill());
         booking.setBookingStatus(BookingStatus.PENDING);
-        booking.setTotalPrice(dto.getTotalPrice());
         booking.setBookingDateTime(LocalDateTime.now());
+        booking.setTotalPrice(days * product.getPrice());
 
         bookingRepository.save(booking);
 
@@ -167,12 +198,12 @@ public class BookingService {
 
         int[] results = new int[days + 1];
 
-        for(Calendar entity: calendar) {
+        for (Calendar entity : calendar) {
             int d = entity.getDate().getDayOfMonth();
 
-            if(entity.isBooked()) {
+            if (entity.isBooked()) {
                 results[d] = 1;
-            } else if(!entity.isBooked()) {
+            } else if (!entity.isBooked()) {
                 results[d] = -1;
             }
         }
@@ -183,10 +214,10 @@ public class BookingService {
         schedule.put("free", new ArrayList<>());
         schedule.put("no booking", new ArrayList<>());
 
-        for(int i = 1; i < results.length; i++) {
-            if(results[i] == 1) {
+        for (int i = 1; i < results.length; i++) {
+            if (results[i] == 1) {
                 schedule.get("booked").add(i);
-            } else if(results[i] == -1) {
+            } else if (results[i] == -1) {
                 schedule.get("free").add(i);
             } else {
                 schedule.get("no booking").add(i);
@@ -202,6 +233,21 @@ public class BookingService {
                         () -> new ResourceNotFoundException("Booking was not found with ID: " + bookingId)
                 );
 
+        User user = getAuthentication();
+
+        if (!booking.getUser().getId().equals(user.getId())) {
+            throw new ProductViolationException("Бронь не принадлежит вам.");
+        }
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
+
+        for (Calendar entity : calendar) {
+            entity.setUser(null);
+            entity.setBooked(false);
+            calendarRepository.save(entity);
+        }
+
         bookingRepository.deleteById(booking.getId());
 
         return "The booking was canceled.";
@@ -213,19 +259,20 @@ public class BookingService {
                         () -> new ResourceNotFoundException("Booking was not found with ID: " + bookingId)
                 );
 
+        if(!getAuthentication().getId().equals(booking.getProduct().getUser().getId())) {
+            throw new ProductViolationException("Вы не можете принять решение по этому продукту, " +
+                    "так как он не принадлежить вам.");
+        }
+
         booking.setBookingStatus(BookingStatus.ACCEPTED);
 
         bookingRepository.save(booking);
 
-        List<Calendar> calendar = booking.getProduct().getCalendars();
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
 
-        for(Calendar entity: calendar) {
-            if((entity.getDate().isEqual(booking.getDateFrom()) ||
-            entity.getDate().isEqual(booking.getDateTill())) ||
-            entity.getDate().isAfter(booking.getDateFrom()) && entity.getDate().isBefore(booking.getDateTill())) {
-                entity.setBooked(true);
-                entity.setUser(booking.getUser());
-            }
+        for (Calendar entity : calendar) {
+            entity.setUser(booking.getUser());
             calendarRepository.save(entity);
         }
 
@@ -238,7 +285,20 @@ public class BookingService {
                         () -> new ResourceNotFoundException("Booking was not found with ID: " + bookingId)
                 );
 
+        if(!getAuthentication().getId().equals(booking.getProduct().getUser().getId())) {
+            throw new ProductViolationException("Вы не можете принять решение по этому продукту, " +
+                    "так как он не принадлежить вам.");
+        }
+
         booking.setBookingStatus(BookingStatus.REJECTED);
+
+        List<Calendar> calendar = calendarRepository.getCalendarByDates(booking.getDateFrom(),
+                booking.getDateTill(), booking.getProduct().getId());
+
+        for (Calendar entity : calendar) {
+            entity.setBooked(false);
+            calendarRepository.save(entity);
+        }
 
         bookingRepository.save(booking);
